+++ strawberryfields/api/result.py
@@ -65,8 +65,11 @@ class Result:
         self._state = None
         self._is_stateful = is_stateful
 
-        # ``samples`` arrives as a list of arrays, need to convert here to a multidimensional array
-        if len(np.shape(samples)) > 1:
+        # samples arrives as either a list of arrays (for shots > 1) or a list (for shots = 1)
+        # need to be converted to a multidimensional array with shape (shots, modes)
+        if np.ndim(samples) == 1:
+            samples = np.array([samples])
+        else:
             samples = np.stack(samples, 1)
         self._samples = samples