+++ strawberryfields/apps/subgraph.py
@@ -39,19 +39,32 @@ Subgraph resizing
 ^^^^^^^^^^^^^^^^^
 
 The key element of the :func:`search` algorithm is the resizing of each subgraph, allowing a
-range of subgraph sizes to be tracked. Resizing proceeds by greedily adding or removing nodes to
-a subgraph one-at-a-time. Node selection is carried out by picking the node with the greatest
-or least degree with respect to to the subgraph, with ties settled uniformly at random.
+range of subgraph sizes to be tracked. Resizing proceeds in the :func:`resize` function by greedily
+adding or removing nodes to a subgraph one-at-a-time. Node selection is carried out by picking
+the node with the greatest or least degree with respect to the subgraph. This function returns a
+dictionary over the range of sizes specified, with each value being the corresponding resized
+subgraph.
 
-This function returns a dictionary over the range of sizes specified, with each value being the
-corresponding resized subgraph.
+Whenever there are multiple candidate nodes with the same degree, there must be a choice of
+which node to add or remove. The supported choices are:
+
+- Select among candidate nodes uniformly at random;
+- Select the candidate node with the greatest node weight, settling remaining ties uniformly at
+  random.
 """
+from typing import Tuple, Union
+
 import networkx as nx
 import numpy as np
 
 
 def search(
-    subgraphs: list, graph: nx.Graph, min_size: int, max_size: int, max_count: int = 10
+    subgraphs: list,
+    graph: nx.Graph,
+    min_size: int,
+    max_size: int,
+    max_count: int = 10,
+    node_select: Union[str, np.ndarray, list] = "uniform",
 ) -> dict:
     """Search for dense subgraphs within an input size range.
 
@@ -59,6 +72,15 @@ def search(
     range specified by ``min_size`` and ``max_size``, resulting in a range of differently sized
     subgraphs. This function loops over all elements of ``subgraphs`` and keeps track of the
     ``max_count`` number of densest subgraphs identified for each size.
+
+    In both growth and shrink phases of :func:`resize`, there may be multiple candidate nodes with
+    equal degree to add to or remove from the subgraph. The method of selecting the node is
+    specified by the ``node_select`` argument, which can be either:
+
+    - ``"uniform"`` (default): choose a node from the candidates uniformly at random;
+    - A list or array: specifying the node weights of the graph, resulting in choosing the node
+      from the candidates with the highest weight (when growing) and lowest weight (when shrinking),
+      settling remaining ties by uniform random choice.
 
     **Example usage:**
 
@@ -80,6 +102,9 @@ def search(
         min_size (int): minimum size to search for dense subgraphs
         max_size (int): maximum size to search for dense subgraphs
         max_count (int): maximum number of densest subgraphs to keep track of for each size
+        node_select (str, list or array): method of settling ties when more than one node of
+            equal degree can be added/removed. Can be ``"uniform"`` (default), or a NumPy array or
+            list containing node weights.
 
     Returns:
         dict[int, list[tuple[float, list[int]]]]: a dictionary of different sizes, each containing a
@@ -89,7 +114,7 @@ def search(
     dense = {}
 
     for s in subgraphs:
-        r = resize(s, graph, min_size, max_size)
+        r = resize(s, graph, min_size, max_size, node_select)
 
         for size, subgraph in r.items():
             r[size] = (nx.density(graph.subgraph(subgraph)), subgraph)
@@ -189,7 +214,13 @@ def _update_subgraphs_list(l: list, t: tuple, max_count: int) -> None:
     return
 
 
-def resize(subgraph: list, graph: nx.Graph, min_size: int, max_size: int) -> dict:
+def resize(
+    subgraph: list,
+    graph: nx.Graph,
+    min_size: int,
+    max_size: int,
+    node_select: Union[str, np.ndarray, list] = "uniform",
+) -> dict:
     """Resize a subgraph to a range of input sizes.
 
     This function uses a greedy approach to iteratively add or remove nodes one at a time to an
@@ -201,9 +232,16 @@ def resize(subgraph: list, graph: nx.Graph, min_size: int, max_size: int) -> dic
     the algorithm performs the procedure again.
 
     When shrinking is required, the algorithm examines all nodes from within the subgraph as
-    candidates and removes the single node with lowest degree relative to the subgraph. In both
-    growth and shrink phases, ties for addition/removal with nodes of equal degree are settled by
-    uniform random choice.
+    candidates and removes the single node with lowest degree relative to the subgraph.
+
+    In both growth and shrink phases, there may be multiple candidate nodes with equal degree to
+    add to or remove from the subgraph. The method of selecting the node is specified by the
+    ``node_select`` argument, which can be either:
+
+    - ``"uniform"`` (default): choose a node from the candidates uniformly at random;
+    - A list or array: specifying the node weights of the graph, resulting in choosing the node
+      from the candidates with the highest weight (when growing) and lowest weight (when shrinking),
+      settling remaining ties by uniform random choice.
 
     **Example usage:**
 
@@ -222,21 +260,16 @@ def resize(subgraph: list, graph: nx.Graph, min_size: int, max_size: int) -> dic
         graph (nx.Graph): the input graph
         min_size (int): minimum size for subgraph to be resized to
         max_size (int): maximum size for subgraph to be resized to
+        node_select (str, list or array): method of settling ties when more than one node of
+            equal degree can be added/removed. Can be ``"uniform"`` (default), or a NumPy array or
+            list containing node weights.
 
     Returns:
         dict[int, list[int]]: a dictionary of different sizes with corresponding subgraph
     """
     nodes = graph.nodes()
     subgraph = set(subgraph)
-
-    if not subgraph.issubset(nodes):
-        raise ValueError("Input is not a valid subgraph")
-    if min_size < 1:
-        raise ValueError("min_size must be at least 1")
-    if max_size >= len(nodes):
-        raise ValueError("max_size must be less than number of nodes in graph")
-    if max_size < min_size:
-        raise ValueError("max_size must not be less than min_size")
+    node_select, w = _validate_inputs(subgraph, graph, min_size, max_size, node_select)
 
     starting_size = len(subgraph)
 
@@ -253,14 +286,19 @@ def resize(subgraph: list, graph: nx.Graph, min_size: int, max_size: int) -> dic
             grow_nodes = grow_subgraph.nodes()
             complement_nodes = nodes - grow_nodes
 
-            degrees = [
-                (c, graph.subgraph(list(grow_nodes) + [c]).degree()[c]) for c in complement_nodes
-            ]
-            np.random.shuffle(degrees)
+            degrees = np.array(
+                [(c, graph.subgraph(list(grow_nodes) + [c]).degree()[c]) for c in complement_nodes]
+            )
+            degrees_max = np.argwhere(degrees[:, 1] == degrees[:, 1].max()).flatten()
 
-            to_add = max(degrees, key=lambda x: x[1])
-            grow_subgraph.add_node(to_add[0])
+            if node_select == "uniform":
+                to_add_index = np.random.choice(degrees_max)
+            elif node_select == "weight":
+                weights = np.array([w[degrees[n][0]] for n in degrees_max])
+                to_add_index = np.random.choice(np.where(weights == weights.max())[0])
 
+            to_add = degrees[to_add_index][0]
+            grow_subgraph.add_node(to_add)
             new_size = grow_subgraph.order()
 
             if min_size <= new_size <= max_size:
@@ -271,11 +309,17 @@ def resize(subgraph: list, graph: nx.Graph, min_size: int, max_size: int) -> dic
         shrink_subgraph = graph.subgraph(subgraph).copy()
 
         while shrink_subgraph.order() > min_size:
-            degrees = list(shrink_subgraph.degree())
-            np.random.shuffle(degrees)
+            degrees = np.array(shrink_subgraph.degree)
+            degrees_min = np.argwhere(degrees[:, 1] == degrees[:, 1].min()).flatten()
 
-            to_remove = min(degrees, key=lambda x: x[1])
-            shrink_subgraph.remove_node(to_remove[0])
+            if node_select == "uniform":
+                to_remove_index = np.random.choice(degrees_min)
+            elif node_select == "weight":
+                weights = np.array([w[degrees[n][0]] for n in degrees_min])
+                to_remove_index = np.random.choice(np.where(weights == weights.min())[0])
+
+            to_remove = degrees[to_remove_index][0]
+            shrink_subgraph.remove_node(to_remove)
 
             new_size = shrink_subgraph.order()
 
@@ -283,3 +327,56 @@ def resize(subgraph: list, graph: nx.Graph, min_size: int, max_size: int) -> dic
                 resized[new_size] = sorted(shrink_subgraph.nodes())
 
     return resized
+
+
+def _validate_inputs(
+    subgraph: set,
+    graph: nx.Graph,
+    min_size: int,
+    max_size: int,
+    node_select: Union[str, np.ndarray, list] = "uniform",
+) -> Tuple:
+    """Validates input for the ``resize`` function.
+
+    This function checks:
+        - if ``subgraph`` is a valid subgraph of ``graph``;
+        - if ``min_size`` and ``max_size`` are sensible numbers;
+        - if ``node_select`` is either ``"uniform"`` or a NumPy array or list;
+        - if, when ``node_select`` is a NumPy array or list, that it is the correct size and that
+          ``node_select`` is changed to ``"weight"``.
+
+    This function returns the updated ``node_select`` and a dictionary mapping nodes to their
+    corresponding weights (weights default to unity if not specified).
+
+    Args:
+        subgraph (list[int]): a subgraph specified by a list of nodes
+        graph (nx.Graph): the input graph
+        min_size (int): minimum size for subgraph to be resized to
+        max_size (int): maximum size for subgraph to be resized to
+        node_select (str, list or array): method of settling ties when more than one node of
+            equal degree can be added/removed. Can be ``"uniform"`` (default), or a NumPy array or
+            list containing node weights.
+
+    Returns:
+        tuple[str, dict]: the updated ``node_select`` and a dictionary of node weights
+    """
+    if not subgraph.issubset(graph.nodes()):
+        raise ValueError("Input is not a valid subgraph")
+    if min_size < 1:
+        raise ValueError("min_size must be at least 1")
+    if max_size >= len(graph.nodes()):
+        raise ValueError("max_size must be less than number of nodes in graph")
+    if max_size < min_size:
+        raise ValueError("max_size must not be less than min_size")
+
+    if isinstance(node_select, (list, np.ndarray)):
+        if len(node_select) != graph.number_of_nodes():
+            raise ValueError("Number of node weights must match number of nodes")
+        w = {n: node_select[i] for i, n in enumerate(graph.nodes)}
+        node_select = "weight"
+    else:
+        w = {n: 1 for i, n in enumerate(graph.nodes)}
+        if node_select != "uniform":
+            raise ValueError("Node selection method not recognized")
+
+    return node_select, w