+++ tests/frontend/test_engine.py
@@ -38,6 +38,19 @@ def prog():
     return prog
 
 
+batch_engines = [
+    sf.Engine("gaussian", backend_options={"batch_size": 2, "cutoff_dim": 6}),
+    sf.Engine("fock", backend_options={"batch_size": 2, "cutoff_dim": 6}),
+    sf.Engine("tf", backend_options={"batch_size": 2, "cutoff_dim": 6}),
+]
+
+engines = [
+    sf.Engine("gaussian", backend_options={"cutoff_dim": 6}),
+    sf.Engine("fock", backend_options={"cutoff_dim": 6}),
+    sf.Engine("tf", backend_options={"cutoff_dim": 6}),
+]
+
+
 class TestEngine:
     """Test basic engine functionality"""